refactor: moved away from BepInEx to full PDX Mods support
