+++ Code/ModSettings.cs
@@ -1,0 +1,34 @@
+﻿// <copyright file="ModSettings.cs" company="ST-Apps (S. Tenuta)">
+// Copyright (c) ST-Apps (S. Tenuta). All rights reserved.
+// Licensed under the Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International License (the "License"); you may not use this file except in compliance with the License.
+// See LICENSE.txt file in the project root for full license information.
+// </copyright>
+
+namespace ExtendedRoadUpgrades.Code
+{
+    using Colossal.IO.AssetDatabase;
+    using Game.Modding;
+
+    /// <summary>
+    /// The mod's settings.
+    /// </summary>
+    [FileLocation(Mod.ModName)]
+    public class ModSettings : ModSetting
+    {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ModSettings"/> class.
+        /// </summary>
+        /// <param name="mod"><see cref="IMod"/> instance.</param>
+        public ModSettings(IMod mod)
+            : base(mod)
+        {
+        }
+
+        /// <summary>
+        /// Restores mod settings to default.
+        /// </summary>
+        public override void SetDefaults()
+        {
+        }
+    }
+}
