--- GlobalProperties.props
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-	<PropertyGroup Label="GlobalProperties">
-		<BepInExVersion>6</BepInExVersion>
-	</PropertyGroup>
-</Project>
