+++ NOTICE.txt
@@ -1,0 +1,3 @@
+Extended Road Upgrades
+/github.com/ST-Apps/CS2-ExtendedRoadUpgrades
+Copyright (c) 2023-24 ST-Apps (S. Tenuta).  All rights reserved.
