+++ Properties/LatestChangelog.md
@@ -1,0 +1,3 @@
+# 1.3.2
+
+- [FIX] Mod not loading properly after v1.3.1
