+++ Properties/LongDescription.md
@@ -1,0 +1,18 @@
+## Description
+This mod enables some additional upgrade modes for your roads, giving you full control on the type of roads you can build without needing to precisely terraform beforehand.
+
+These upgrades are available in the in-game upgrade mode for roads and can be used exactly like any other upgrade mode (e.g. Grass or Wide Sidewalk).
+
+The mod currentyl supports the following modes:
+- Quay
+- Retaining Wall
+- Elevated
+- Tunnel
+
+These upgrades can be applied to:
+- Roads
+- Pedestrian Paths
+- Public Transport Roads
+- Train Tracks
+- Tram Tracks
+- Subway Tracks
