+++ l10n/en-US.csv
@@ -1,0 +1,8 @@
+Assets.NAME[Quay]	Quay
+Assets.DESCRIPTION[Quay]	"A quay, if you installed this mod you know what it is :)"
+Assets.NAME[RetainingWall]	Retaining Wall
+Assets.DESCRIPTION[RetainingWall]	"A retaining wall, if you installed this mod you know what it is :)"
+Assets.NAME[Elevated]	Elevated
+Assets.DESCRIPTION[Elevated]	"Elevated mode, kind of similar to bridges."
+Assets.NAME[Tunnel]	Tunnel
+Assets.DESCRIPTION[Tunnel]	"Tunnel mode, it might not look perfect but it works."
