+++ l10n/it-IT.csv
@@ -1,0 +1,8 @@
+Assets.NAME[Quay]	Banchina
+Assets.DESCRIPTION[Quay]	"Una banchina, se hai installato la mod sai di cosa si tratta :)"
+Assets.NAME[RetainingWall]	Muro di contenimento
+Assets.DESCRIPTION[RetainingWall]	"Un muro di contenimento, se hai installato la mod sai di cosa si tratta :)"
+Assets.NAME[Elevated]	Sopraelevata
+Assets.DESCRIPTION[Elevated]	"Modalità sopraelevata, simile ad un ponte."
+Assets.NAME[Tunnel]	Tunnel
+Assets.DESCRIPTION[Tunnel]	"Modalità tunnel, non è perfetta ma funziona."
