--- manifest.json.template
@@ -1,9 +1,0 @@
-{
-  "name": "Extended_Road_Upgrades",
-  "version_number": "{ProjectVersion}",
-  "website_url": "https://github.com/ST-Apps/CS2-ExtendedRoadUpgrades",
-  "description": "A mod that enables multiple upgrade options for your roads.",
-  "dependencies": [
-    "BepInEx-BepInExPack-5.4.2100"
-  ]
-}
